Convert CXFA_Node::GetChild to take a size_t for the index

+++ xfa/fxfa/parser/cxfa_node.h
@@ -157,7 +157,7 @@ class CXFA_Node : public CXFA_Object {
   size_t CountChildren(XFA_Element eType, bool bOnlyChild);
 
   template <typename T>
-  T* GetChild(int32_t index, XFA_Element eType, bool bOnlyChild) {
+  T* GetChild(size_t index, XFA_Element eType, bool bOnlyChild) {
     return static_cast<T*>(GetChildInternal(index, eType, bOnlyChild));
   }
 
@@ -247,9 +247,7 @@ class CXFA_Node : public CXFA_Object {
   void OnRemoved(bool bNotify);
   pdfium::Optional<void*> GetDefaultValue(XFA_Attribute attr,
                                           XFA_AttributeType eType) const;
-  CXFA_Node* GetChildInternal(int32_t index,
-                              XFA_Element eType,
-                              bool bOnlyChild);
+  CXFA_Node* GetChildInternal(size_t index, XFA_Element eType, bool bOnlyChild);
   CXFA_Node* GetFirstChildByClassInternal(XFA_Element eType) const;
   CXFA_Node* GetNextSameNameSiblingInternal(
       const WideStringView& wsNodeName) const;
